ゴールデンパスでパラメータ推定が動くようになった。
git-svn-id: file:///home/moriyoshi/chaperone.e-cell.org/var/opt/subversion/ecell-ide/trunk@1017 fe1fbaf7-ad0a-4b3d-a902-8cb9c273ac6b

+++ ide/Analysis/Analysis/PEAdvancedWindow.cs
@@ -71,7 +71,7 @@ namespace EcellLib.Analysis
         public SimplexCrossoverParameter GetParam()
         {
             return new SimplexCrossoverParameter(
-                Convert.ToDouble(PEMTextBox.Text),
+                Convert.ToInt32(PEMTextBox.Text),
                 Convert.ToDouble(PEM0TextBox.Text),
                 Convert.ToDouble(PEMaxRateTextBox.Text),
                 Convert.ToDouble(PEKTextBox.Text),
